multipleEcommerce/multipleEcommerce/product/models.py
from django.db import models
from mptt.models import MPTTModel, TreeForeignKey

# Create your models here, they are going to be used to create a file of what is
#  expected to be created at the database.


class Category(MPTTModel):
    name = models.CharField(max_length=100, unique=True)
    parent = TreeForeignKey("self", on_delete=models.PROTECT, null=True, blank=True)

    class MPTTMeta:
        order_insertion_by = ["name"]

    def __str__(self):
        return self.name


class Brand(models.Model):
    name = models.CharField(max_length=100)

    # this use the name used in the input in the front end
    # if you don't use this, the name will be object 1, object 2, etc
    def __str__(self):
        return self.name


class Product(models.Model):
    name = models.CharField(max_length=100)
    description = models.TextField(blank=True)
    brand = models.ForeignKey(Brand, on_delete=models.CASCADE)
    category = TreeForeignKey("Category", on_delete=models.SET_NULL, null=True, blank=True)
    is_active = models.BooleanField(default=False)

    def __str__(self):
        return self.name


class ProductDetails(models.Model):
    product = models.ForeignKey(Product, on_delete=models.CASCADE)
    is_active = models.BooleanField(default=False)
    contact = models.TextField(blank=True)
    quantity = models.IntegerField()

    def __str__(self):
        return self.product.name
